+++ idtx_sale_order/models/product_template.py
@@ -7,7 +7,7 @@ class ProductTemplate(models.Model):
     analysis_product_family_id = fields.Many2one(related='analysis_id.product_family_id', store=True, readonly=True, index=True)
     product_title_id = fields.Many2one(related='analysis_id.product_title_id', store=True, readonly=True, index=True)
     density = fields.Integer(related='analysis_id.density', store=True, readonly=True, index=True)
-    standard_width = fields.Float(related='analysis_id.standard_width', store=True, readonly=True, index=True)
+    standard_width = fields.Integer(related='analysis_id.standard_width', store=True, readonly=True, index=True)
 
     @api.depends('analysis_id')
     def _compute_is_rect(self):
